Create TextFieldControl.java
This is a demonstration, to have control over the text field, which will disable the direct copy past options (Ctrl + C and Ctrl + V). This ensures the security which disabling a copy paste option when the user opens the book, as it is going to open in a text area (controls for text area and text field works almost the same).